refactor(RecordServiceImpl, RecordSerivce): 기록 서비스 로직 인터페이스 구현체 구현
